Changed a lot of stuff
changed the way objects holds its map. Its now a map of bools to []objects.

updated bat.Moved() to only save newBox as saving pos was not needed.

ignore all commented shit in game.go. used for testing hitbox stuff.

fixed some hitbox bugs in object.go as i had done the moved() wrong. Also added schale as a value to Draw so you can set the scale value for objects.

Added more to world.go. Starting to cteate world for bat to be in. Starter area is going to be a cave.