Update alpha 0.1.1
Update Notes:

Removed code for hit box colition. This will be getting a overhall soon.

Removed some of the old world code as it got reworked. The primary change for this update was implimenting loading from a Tiled json file and loading the map to the screen.

Removed all tilepix code as it did not work correctly and was missing large portions of importent fetures. Instead I implemented the basic fetures my self. All files int he mapdecoder pacages is part of the json decoding for Tilepix. The world.go file holds the ocncersion from the basic struct types form the json files to a usible format to display on the screen.