Update Alpha 0.1.3
Large Scale Refactoring and optimizations.

Nearly every file had some code changed. The way actions and animations are handled are now completly difforent. Moved animations to be part of bat and changed code to fallow a Update() Draw() structure for all main opjects.

Fixed most of the odd rendering issues with the map by nolonger clearing the batch for a layer unless that layer has been changed in some way.